test: rename withVirtualTestnet to createVirtualTest and remove tenderly account config from barrel
