+++ tests/cypress/support/helpers/tenderly/index.ts
@@ -1,15 +1,1 @@
-export * from './types'
-export {
-  createVirtualTestnet,
-  type CreateVirtualTestnetOptions,
-  type CreateVirtualTestnetResponse,
-} from './vnet-create'
-export { deleteVirtualTestnet, type DeleteVirtualTestnetOptions } from './vnet-delete'
-export { withVirtualTestnet, createTestWagmiConfigFromVNet } from './vnet'
-
-const accountSlug = Cypress.env('TENDERLY_ACCOUNT_SLUG')
-const projectSlug = Cypress.env('TENDERLY_PROJECT_SLUG')
-const accessKey = Cypress.env('TENDERLY_ACCESS_KEY')
-
-export const tenderlyAccount =
-  accountSlug && projectSlug && accessKey ? { accountSlug, projectSlug, accessKey } : undefined
+export { createVirtualTestnet, createTestWagmiConfigFromVNet } from './vnet'
