--- tests/cypress/support/helpers/tenderly/types.ts
@@ -1,5 +1,0 @@
-export type TenderlyAccount = {
-  accountSlug: string
-  projectSlug: string
-  accessKey: string
-}
